feature/ATC-219 - Adds babel-server gulp task * Removes unused assets tasks and unused vars in paths.js * Moves ./index.js to src/scripts/server/index.js * Adds new npm script to start node server * Updates Procfile to use new npm commands
