+++ package.json
@@ -13,7 +13,8 @@
     "report": "nyc report --reporter=cobertura  --reporter=lcov",
     "coverage": "npm run test && npm run report",
     "lint": "eslint src/**; exit 0",
-    "start": "gulp server"
+    "start": "gulp server",
+    "server": "node assets/scripts/server"
   },
   "repository": {
     "type": "git",
@@ -67,6 +68,7 @@
     ]
   },
   "dependencies": {
+    "bufferutil": "^1.3.0",
     "chalk": "^1.1.3",
     "express": "^4.14.0",
     "jquery": "^3.1.0",
@@ -75,6 +77,7 @@
     "raf": "^3.3.0",
     "socket.io": "^1.7.2",
     "socket.io-client": "^1.7.2",
+    "utf-8-validate": "^2.0.0",
     "uuid4": "^1.0.0"
   },
   "devDependencies": {
@@ -97,13 +100,13 @@
     "eslint-plugin-react": "^6.2.0",
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.1",
+    "gulp-babel": "^6.1.2",
     "gulp-cache": "^0.4.5",
     "gulp-clean": "^0.3.2",
     "gulp-cli": "^1.2.2",
     "gulp-concat": "^2.6.0",
     "gulp-connect": "^5.0.0",
     "gulp-eslint": "^3.0.1",
-    "gulp-imagemin": "^3.0.3",
     "gulp-minify-css": "^1.2.4",
     "gulp-rename": "^1.2.2",
     "gulp-sourcemaps": "^1.6.0",
