+++ tools/paths.js
@@ -8,21 +8,17 @@ var src = './src';
 var srcAssetsDir = path.join(src, 'assets');
 var scrScriptsFonts = path.join(srcAssetsDir, 'fonts');
 var srcScriptsDir = path.join(srcAssetsDir, 'scripts');
+var srcScriptsServerDir = path.join(srcAssetsDir, 'scripts/server');
 var srcStylesDir = path.join(srcAssetsDir, 'style');
 var srcImagesDir = path.join(srcAssetsDir, 'images');
-var srcAircraft = path.join(srcAssetsDir, 'aircraft');
-var srcAirlines = path.join(srcAssetsDir, 'airlines');
-var srcAirports = path.join(srcAssetsDir, 'airports');
 
 var build = ROOT;
 var buildAssetsDir = path.join(ROOT, 'assets');
 var buildFontsDir = path.join(buildAssetsDir, 'fonts');
 var buildScriptsDir = path.join(buildAssetsDir, 'scripts');
+var buildScriptsServerDir = path.join(buildAssetsDir, 'scripts/server');
 var buildStylesDir = path.join(buildAssetsDir, 'style');
 var buildImagesDir = path.join(buildAssetsDir, 'images');
-var buildAircraft = path.join(buildAssetsDir, 'aircraft');
-var buildAirlines = path.join(buildAssetsDir, 'airlines');
-var buildAirports = path.join(buildAssetsDir, 'airports');
 
 // configuration object that wraps all path/file/glob
 var options = {};
@@ -32,26 +28,24 @@ options.DIR = {
     SRC_ASSETS: srcAssetsDir,
     SRC_FONTS: scrScriptsFonts,
     SRC_SCRIPTS: srcScriptsDir,
+    SRC_SCRIPTS_SERVER: srcScriptsServerDir,
     SRC_STYLE: srcStylesDir,
     SRC_IMAGES: srcImagesDir,
-    SRC_AIRCRAFT: srcAircraft,
-    SRC_AIRLINES: srcAirlines,
-    SRC_AIRPORTS: srcAirports,
 
     BUILD: build,
     BUILD_ASSETS: buildAssetsDir,
     BUILD_FONTS: buildFontsDir,
     BUILD_SCRIPTS: buildScriptsDir,
+    BUILD_SCRIPTS_SERVER: buildScriptsServerDir,
     BUILD_STYLE: buildStylesDir,
     BUILD_IMAGES: buildImagesDir,
-    BUILD_AIRCRAFT: buildAircraft,
-    BUILD_AIRLINES: buildAirlines,
-    BUILD_AIRPORTS: buildAirports,
 };
 
 // for specific files
 options.FILE = {
-    JS_ENTRY: path.join(options.DIR.SRC_SCRIPTS, 'index.js'),
+    JS_ENTRY_CLIENT: path.join(options.DIR.SRC_SCRIPTS, 'index.js'),
+    // TODO: this may need to move to a const
+    JS_ENTRY_SERVER: path.join(options.DIR.SRC_SCRIPTS_SERVER, 'index.js'),
     CSS_MAIN: path.join(options.DIR.SRC_STYLE, 'main.css'),
     PKG_JSON: path.join(options.ROOT, 'package.json')
 };
